+++ main.tf
@@ -3,7 +3,7 @@ resource "google_service_account_iam_binding" "binding" {
 
   service_account_id = var.service_account_id
   role               = var.role
-  members            = var.members
+  members            = [for m in var.members : try(var.computed_members_map[regex("^computed:(.*)", m)[0]], m)]
 
   depends_on = [var.module_depends_on]
 }
@@ -13,7 +13,7 @@ resource "google_service_account_iam_member" "member" {
 
   service_account_id = var.service_account_id
   role               = var.role
-  member             = each.value
+  member             = try(var.computed_members_map[regex("^computed:(.*)", each.value)[0]], each.value)
 
   depends_on = [var.module_depends_on]
 }
@@ -22,7 +22,7 @@ resource "google_service_account_iam_policy" "policy" {
   count = var.module_enabled && var.policy_bindings != null ? 1 : 0
 
   service_account_id = var.service_account_id
-  policy_data        = data.google_iam_policy.policy[0].policy_data
+  policy_data        = try(data.google_iam_policy.policy[0].policy_data, null)
 
   depends_on = [var.module_depends_on]
 }
@@ -35,7 +35,7 @@ data "google_iam_policy" "policy" {
 
     content {
       role    = binding.value.role
-      members = try(binding.value.members, var.members)
+      members = [for m in binding.value.members : try(var.computed_members_map[regex("^computed:(.*)", m)[0]], m)]
 
       dynamic "condition" {
         for_each = try([binding.value.condition], [])