+++ versions.tf
@@ -1,10 +1,10 @@
 terraform {
-  required_version = ">= 0.14, < 2.0"
+  required_version = "~> 1.0"
 
   required_providers {
     google = {
       source  = "hashicorp/google"
-      version = ">= 3.59, < 5.0"
+      version = "~> 4.0"
     }
   }
 }